#94 fix rounding issue in total frames for output window + testcases
