+++ goddo_test/test_preview_output_playback.py
@@ -27,430 +27,6 @@ def test_slider_disabled_initially(windows_container: WindowsContainer):
     assert windows_container.output_window.preview_widget.frame_pixmap is None
     assert windows_container.output_window.slider.value() == 0
 
-def test_restricted_keyboard_move_to_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:00.01')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.12, 0.13), current_frame_no=expected_in_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_try_to_move_to_before_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:00.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    #   try move to before in frame
-    pyautogui.press('left')
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.12, 0.13), current_frame_no=expected_in_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_move_right_from_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:00.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = expected_in_frame + 1
-
-    #   from in frame move right
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no ==expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.13, 0.14), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_move_left_from_bet_in_and_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:30.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 90 - 5
-
-    #   from in frame move right
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.48, 0.49), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_move_right_from_bet_in_and_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:30.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 90 + 1
-
-    #   from in frame move right
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.49, 0.51), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_move_to_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:59.03')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.87, 0.88), current_frame_no=expected_out_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_try_to_move_pass_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:00.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    pyautogui.press('right')
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.87, 0.88), current_frame_no=expected_out_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_keyboard_move_left_from_out_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:00.00')
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = expected_out_frame - 5
-
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.85, 0.86), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_keyboard_move_to_start_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 50
-
-    #   move to start frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:50.01')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0, 0.01), current_frame_no=expected_cur_frame, is_output_window=True, restrict_frame_interval=False, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_keyboard_try_to_move_pass_start_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 50
-
-    #   move to start frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:50.00')
-    assert app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame
-
-    pyautogui.press('left')
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0, 0.01), current_frame_no=expected_cur_frame, is_output_window=True, restrict_frame_interval=False, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_keyboard_move_left_bet_start_and_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 55 - 5
-
-    #   move to between start and in frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:55.00')
-
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.04, 0.05), current_frame_no=expected_cur_frame, is_output_window=True, restrict_frame_interval=False, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_keyboard_move_right_bet_start_and_in_frame(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_cur_frame = 4 * 55 + 1
-
-    #   move to between start and in frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:55.00')
-
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_cur_frame)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.06, 0.07), current_frame_no=expected_cur_frame, is_output_window=True, restrict_frame_interval=False, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_unrestricted_move_with_keyboard(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    pyautogui.press('f')
-    wait_until(lambda: not app_thread.mon.state.preview_window_output.restrict_frame_interval)
-
-    expected_start_frame = 4 * 50
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-    expected_end_frame = 4 * (60 * 2 + 10)
-
-    #   move to start frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:50.01')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame)
-
-    #   try move to before start frame
-    pyautogui.press('left')
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame)
-    
-    #   from start frame move right
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_start_frame + 1)
-
-    #   in between start and in frame, move left and right 
-    enter_time_in_go_to_dialog_box(app_thread, '0:00:55.00')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 55 - 5)
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 55 - 4)
-
-    #   go back and forth on in frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:01.00')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 60 - 1)
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 60)
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 60 + 1)
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:00.00')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 60 - 5)
-
-    #   between in frame and out frame, move left and right
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:30.00')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 90 - 5)
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 90 - 4)
-
-    #   go back and forth on out frame
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:59.03')
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame + 1)
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame - 4)
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:00.00')
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame - 5)
-
-    #   in between out frame and end frame, move left and right 
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:05.00')
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 125 + 1)
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 125 - 4)
-
-    #   move to end frame, 
-    enter_time_in_go_to_dialog_box(app_thread, '0:02:09.03')
-    pyautogui.press('right')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame)
-
-    pyautogui.press('right')
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame)
-
-    pyautogui.press('left')
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_end_frame - 5)
-
-    expected_cur_frame = expected_end_frame - 5
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.975, 0.985), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
-def test_restricted_move_with_mouse_scroll(app_thread, windows_container: WindowsContainer, blank_state):
-    open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
-
-    expected_in_frame = 4 * 60 * 1
-    expected_out_frame = 4 * 60 * 2
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:01.00')
-    go_to_prev_wind_slider(windows_container.output_window, 0.1)
-
-    # try to scroll pass in frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-
-    # when exactly on in frame, try to scroll back
-    pyautogui.scroll(1)
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-
-    # scroll forward when on in frame
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 65)
-
-    # scroll forward when between in and out frame
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 70)
-
-    # scroll backward when between in and out frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 65)
-
-    # scroll back to in frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_in_frame)
-
-    enter_time_in_go_to_dialog_box(app_thread, '0:01:59.00')
-
-    # try to scroll pass out frame
-    pyautogui.scroll(-1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-
-    # when exactly on out frame, try to scroll forward
-    pyautogui.scroll(-1)
-    time.sleep(0.5)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == expected_out_frame)
-
-    # scroll backward when on out frame
-    pyautogui.scroll(1)
-    wait_until(lambda: app_thread.mon.state.preview_window_output.current_frame_no == 4 * 115)
-
-    clip = get_video_clip_for_1hr_vid(in_frame=expected_in_frame, out_frame=expected_out_frame)
-    expected_timeline_clips = [(clip,120)]
-
-    expected_cur_frame = 4 * 115
-
-    generic_assert(app_thread, windows_container, blank_state,
-                get_assert_file_list_for_1hr_fn(), get_assert_blank_list_fn(is_file_list=False), 
-                get_assert_preview_fn(clip, slider_range=(0.03, 0.04), current_frame_no=expected_out_frame),
-                get_assert_preview_fn(clip, slider_range=(0.81, 0.82), current_frame_no=expected_cur_frame, is_output_window=True, extra_frames_left=40, extra_frames_right=40), 
-                get_assert_timeline_fn(expected_timeline_clips, selected_clip_index=0, opened_clip_index=0))
-
 def test_unrestricted_move_with_mouse_scroll(app_thread, windows_container: WindowsContainer, blank_state):
     open_clip_on_output_window(app_thread, windows_container, '0:01:00.00', '0:02:00.00', get_blank_1hr_vid_path())
