+++ goddo_test/utils/assert_utils.py
@@ -434,8 +434,8 @@ def get_assert_preview_fn(clip: VideoClip, slider_range=(0.00, 0.01), current_fr
             cur_abs_time_label, cur_rel_time_label, total_time_label, speed_label, skip_label, restrict_label, time_color = \
                 parse_preview_window_label(resolved_win_state_dict['label'], WINDOW_NAME_OUTPUT)
             defaulted_restrict_label = f'restrict={defaulted_restrict_frame_interval if restrict_frame_interval is None else restrict_frame_interval}'
-            clip_total_frames = clip.frame_in_out.get_no_of_frames(clip.total_frames)
-            cur_total_frames = clip_total_frames + extra_frames_left + extra_frames_right
+            clip_total_frames = clip.frame_in_out.get_no_of_frames(clip.total_frames) - 1
+            cur_total_frames = clip.frame_in_out.get_no_of_frames(clip.total_frames) + extra_frames_left + extra_frames_right
             cur_start_frame = clip.frame_in_out.get_resolved_in_frame() - extra_frames_left
             cur_end_frame = clip.frame_in_out.get_resolved_out_frame(clip.total_frames) + extra_frames_right
             expected_total_time_label = clip.get_total_time_str(clip_total_frames)
@@ -453,9 +453,10 @@ def get_assert_preview_fn(clip: VideoClip, slider_range=(0.00, 0.01), current_fr
             cur_abs_time_label, cur_rel_time_label, total_time_label, speed_label, skip_label, restrict_label, time_color = \
                 parse_preview_window_label(resolved_win_state_dict['label'], WINDOW_NAME_SOURCE)
             defaulted_restrict_label = '' if restrict_label is None else restrict_label
-            cur_total_frames = clip.total_frames
+            clip_total_frames = clip.total_frames
+            cur_total_frames = clip_total_frames
             cur_start_frame = 1
-            cur_end_frame = clip.total_frames
+            cur_end_frame = clip_total_frames
             expected_total_time_label = clip.get_total_time_str()
 
         # preview window asserts
@@ -499,6 +500,9 @@ def get_assert_preview_fn(clip: VideoClip, slider_range=(0.00, 0.01), current_fr
             cur_rel_frames_in_time_label = time_str_to_frames(cur_rel_time_label, clip.fps)
             cur_total_frames_in_time_label = time_str_to_frames(total_time_label, clip.fps)
             assert cur_abs_frames_in_time_label == (cur_rel_frames_in_time_label + clip.frame_in_out.get_resolved_in_frame())
+
+            logging.info(f'cur_rel_time_label {cur_rel_time_label} total_time_label {total_time_label}')
+            logging.info(f'=== cur rel frames {cur_rel_frames_in_time_label} cur total frames {cur_total_frames_in_time_label} cur frame {resolved_state_dict["current_frame_no"]} resolve in frame {clip.frame_in_out.get_resolved_in_frame()} resolve out frame {clip.frame_in_out.get_resolved_out_frame(clip.total_frames)} clip total frames {clip.total_frames}')
 
             if resolved_state_dict['current_frame_no'] < clip.frame_in_out.get_resolved_in_frame():
                 assert cur_rel_frames_in_time_label < 0
